{{ cookiecutter.project_slug }}/requirements/base.txt
# Core
pytz==2022.1
Django==4.0.4
gunicorn==20.1.0
newrelic==7.6.0.173
Pillow==9.1.0
django_cors_headers==3.12.0
django-environ==0.8.1

# For the persistence stores
psycopg2-binary==2.9.3
dj-database-url==0.5.0

# Model Tools
django-model-utils==4.2.0
django_unique_upload==0.2.1
django-extensions==3.1.5
django_simple_history==3.0.0
django-phonenumber-field==6.1.0
phonenumbers==8.12.48

# Rest apis
djangorestframework==3.13.1
Markdown==3.3.7
django-filter==21.1
djoser==2.1.0
djangorestframework-camel-case==1.3.0
easy_thumbnails==2.8.1
drf-extensions==0.7.1

# Task Queue
dramatiq[redis, watch]==1.13.0
django-dramatiq==0.11.0
hiredis==2.0.0

# Static and Media Storage
django-storages==1.12.3
boto3==1.21.13

# Health check for AWS
django-health-check==3.16.5
